update materials changes
