+++ src/components/BMDashboard/MaterialsList/RecordsModal.jsx
@@ -58,6 +58,7 @@ export function Record({ record, recordType }) {
   //     </>
   //   );
   // }
+
   if (recordType === 'Update') {
     return (
       <>
@@ -70,15 +71,15 @@ export function Record({ record, recordType }) {
           </tr>
         </thead>
         <tbody>
-          {record.updateRecord.map(({ date, quantityUsed, quantityWasted, createdBy }) => {
+          {record.updateRecord.map((data, idx) => {
             return (
-              <tr key={date + createdBy._id}>
-                <td>{moment(date).format('MM/DD/YY')}</td>
-                <td>{quantityUsed + ' ' + record.itemType?.unit || '-'}</td>
-                <td>{quantityWasted + ' ' + record.itemType?.unit || '-'}</td>
+              <tr key={idx}>
+                <td>{moment.utc(data.date).format('LL')}</td>
+                <td>{data.quantityUsed + ' ' + record.itemType?.unit || '-'}</td>
+                <td>{data.quantityWasted + ' ' + record.itemType?.unit || '-'}</td>
                 <td>
-                  <a href={`/userprofile/${createdBy._id}`}>
-                    {`${createdBy.firstName} ${createdBy.lastName}`}
+                  <a href={`/userprofile/${data.createdBy._id}`}>
+                    {`${data.createdBy.firstName} ${data.createdBy.lastName}`}
                   </a>
                 </td>
               </tr>
